refactor(rpc): extract row codec helpers
Move row/key byte encoding and decoding helpers out of lineairdb_rpc.cc
into row_codec.{cc,hh}. Register row_codec.cc in the server build.
